move manySegmentsTest to ParallelMonitorSolrQueryTest
